+++ CHANGELOG.md
@@ -16,6 +16,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - CLI: fix the display of the assets in yaml and json format (#115)
 - No more crash when running a script in local subprocess mode from a path with spaces (#99)
 
+### Added
+
+- Error type to traintuple, testtuple, aggregatetuple and composite_traintuple
+
 ## [0.14.0] - 2021-12-01
 
 ### Added